Moving generic stuff out of deequ module
